Ident (ident v0.1.0)

This is a method that creates a unique Identicon for an inputed string and outputs a png file with the unique Identicon image.

Functions

create_grid(image)

Specs

create_grid(%Ident.Image{
  color: term(),
  grid: any(),
  grid_squares: term(),
  hex: list()
}) :: %Ident.Image{
  color: term(),
  grid: [{any(), integer()}],
  grid_squares: term(),
  hex: list()
}

This function creates a list of tuples containing the hex number and its index.

Function

def create_grid(%Ident.Image{hex: hex} = image) do
  grid =
    hex
    |> List.delete_at(length(hex) - 1)
    |> Enum.chunk_every(3)
    |> Enum.map(&create_row/1)
    |> List.flatten()
    |> Enum.with_index()

  %Ident.Image{image | grid: grid}
end

create_row(chunk)

Specs

create_row([...]) :: [...]

This functions creates a chunk or sub-list that is symetrical.

Function

def create_row(chunk) do
  [a, b, _c] = chunk
  chunk ++ [b, a]
end

create_squares(image)

Specs

create_squares(%Ident.Image{
  color: term(),
  grid: any(),
  grid_squares: any(),
  hex: term()
}) :: %Ident.Image{
  color: term(),
  grid: any(),
  grid_squares: list(),
  hex: term()
}

This functions creates the points for the squares. It creates a list of tuples of tuples that contains the coordinates of the top left square and the coordinate of the bottom right square. Calculated using the index.

Function

def create_squares(%Ident.Image{grid: grid} = image) do
  grid_squares =
    Enum.map(grid, fn {_code, index} ->
      horizontal = rem(index, 5) * 50
      vertical = div(index, 5) * 50

      top_left = {horizontal, vertical}
      bottom_right = {horizontal + 50, vertical + 50}

      {top_left, bottom_right}
    end)

  %Ident.Image{image | grid_squares: grid_squares}
end

draw_the_image(image)

Specs

draw_the_image(%Ident.Image{
  color: atom() | {byte(), byte(), byte()} | {byte(), byte(), byte(), byte()},
  grid: term(),
  grid_squares: any(),
  hex: term()
}) :: binary()

This function draws the image with the egd drawer and then renders the image to container built.

Function

def draw_the_image(%Ident.Image{color: color, grid_squares: grid_squares}) do
  container = :egd.create(250, 250)
  color_square = :egd.color(color)

  Enum.each(grid_squares, fn {top_left, bottom_right} ->
    :egd.filledRectangle(container, top_left, bottom_right, color_square)
  end)

  :egd.render(container)
end

filter_odds(image)

Specs

filter_odds(%Ident.Image{
  color: term(),
  grid: any(),
  grid_squares: term(),
  hex: term()
}) :: %Ident.Image{
  color: term(),
  grid: list(),
  grid_squares: term(),
  hex: term()
}

This function filters the grid to only output even hex numbers.

Function

def filter_odds(%Ident.Image{grid: grid} = image) do
  filtered_grid =
    Enum.filter(grid, fn {code, _index} ->
      rem(code, 2) == 0
    end)

  %Ident.Image{image | grid: filtered_grid}
end

hash_string(string)

Specs

hash_string(
  binary()
  | maybe_improper_list(
      binary() | maybe_improper_list(any(), binary() | []) | byte(),
      binary() | []
    )
) :: any()

This function turns a string into a hash with the md5 algorithm. It then converts that to a list of integers.

Function

def hash_string(string) do
  hex =
    :crypto.hash(:md5, string)
    |> :binary.bin_to_list()

  %Ident.Image{hex: hex}
end

main(string, filename)

Specs

main(
  binary()
  | maybe_improper_list(
      binary() | maybe_improper_list(any(), binary() | []) | byte(),
      binary() | []
    ),
  any()
) :: :ok | {:error, atom()}

This is the main function that takes a string and filename as arguments. Then outputs the identicon to a png file.

Function

def main(string, filename) do
  string
  |> hash_string
  |> pick_color
  |> create_grid
  |> filter_odds
  |> create_squares
  |> draw_the_image
  |> save_image(filename)
end

pick_color(image)

Specs

pick_color(%Ident.Image{
  color: any(),
  grid: term(),
  grid_squares: term(),
  hex: nonempty_maybe_improper_list()
}) :: %Ident.Image{
  color: {any(), any(), any()},
  grid: term(),
  grid_squares: term(),
  hex: nonempty_maybe_improper_list()
}

This function picks the color from the first three integers in the list.

Function

def pick_color(%Ident.Image{hex: [r, g, b | _tail]} = image) do
  %Ident.Image{image | color: {r, g, b}}
end

save_image(image, filename)

Specs

save_image(
  binary()
  | maybe_improper_list(
      binary() | maybe_improper_list(any(), binary() | []) | byte(),
      binary() | []
    ),
  any()
) :: :ok | {:error, atom()}

This function saves the final image to the the root of the directory with the filename inputed.

##Function

def save_image(image, filename) do
  File.write("#{filename}.png", image)
end